Corrected auth module routing and added login and reminder routes
