+++ src/store/Auth/actions.js
@@ -1,5 +1,5 @@
 /* {[The file is published on the basis of YetiForce Public License 3.0 that can be found in the following directory: licenses/LicenseEN.txt or yetiforce.com]} */
-import loginAxios from '../../services/Login.js'
+import loginAxios from '../../services/Auth.js'
 import globalAxios from '../../services/Global.js'
 import { LocalStorage } from 'quasar'
 import actions from '../actions.js'
@@ -13,13 +13,17 @@ export default {
    * @param   {any}     dispatch
    * @param   {object}  user
    */
-  [actions.Login.login]({ commit, dispatch }, user) {
+  [actions.Auth.login]({ commit, dispatch }, user) {
     loginAxios({
       url: 'login.php/',
       data: user,
       method: 'POST'
     })
-      .then(({ data }) => {
+      .then(response => {
+        const data = response.data
+        if (response.status !== 200) {
+          return console.error('Server error', response)
+        }
         const now = new Date()
         const expirationDate = new Date(now.getTime() + data.expiresIn * 100000)
         LocalStorage.set('tokenId', data.tokenId)
@@ -27,14 +31,14 @@ export default {
         LocalStorage.set('userName', data.userName)
         LocalStorage.set('admin', data.admin)
         LocalStorage.set('expiresIn', expirationDate)
-        commit(mutations.Login.authUser, {
+        commit(mutations.Auth.authUser, {
           tokenId: data.tokenId,
           userId: data.userId,
           admin: data.admin,
           userName: data.userName
         })
         globalAxios.defaults.headers.common['Authorization'] = data.tokenId
-        dispatch(actions.Login.setLogoutTimer, data.expiresIn)
+        dispatch(actions.Auth.setLogoutTimer, data.expiresIn)
         this.$router.replace('/')
       })
       .catch(error => console.log(error))
@@ -51,9 +55,9 @@ export default {
    * @param   {any}     dispatch
    * @param   {number}  expirationTime
    */
-  [actions.Login.setLogoutTimer]({ commit }, expirationTime) {
+  [actions.Auth.setLogoutTimer]({ commit }, expirationTime) {
     setTimeout(() => {
-      commit(mutations.Login.clearAuthData)
+      commit(mutations.Auth.clearAuthData)
     }, expirationTime * 100000)
   },
 
@@ -62,16 +66,16 @@ export default {
    *
    * @param   {any}     commit
    */
-  [actions.Login.tryAutoLogin]({ commit }) {
+  [actions.Auth.tryAutoLogin]({ commit }) {
     return new Promise(resolve => {
       const token = localStorage.getItem('tokenId')
       const expirationDate = new Date(localStorage.getItem('expiresIn')).getTime()
       const now = new Date().getTime()
       if (!token || now >= expirationDate) {
-        commit(mutations.Login.clearAuthData)
+        commit(mutations.Auth.clearAuthData)
         resolve(false)
       } else {
-        commit(mutations.Login.authUser, {
+        commit(mutations.Auth.authUser, {
           tokenId: token,
           userId: localStorage.getItem('userId'),
           admin: localStorage.getItem('admin'),
