+++ src/store/actions.js
@@ -1,8 +1,8 @@
 /* {[The file is published on the basis of YetiForce Public License 3.0 that can be found in the following directory: licenses/LicenseEN.txt or yetiforce.com]} */
 export default {
-  Login: {
-    login: 'Login/login',
-    setLogoutTimer: 'Login/setLogoutTimer',
-    tryAutoLogin: 'Login/tryAutoLogin'
+  Auth: {
+    login: 'Auth/login',
+    setLogoutTimer: 'Auth/setLogoutTimer',
+    tryAutoLogin: 'Auth/tryAutoLogin'
   }
 }
